defer cosmetic filter key calculation to engine-add time
the info is essentially an engine-internal optimization detail and won't
be useful for external users of the library
